requirements.txt
iniconfig==2.0.0
mysql-connector-python==9.1.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.2
